fix post merge errors

+++ src/Etterna/Singletons/NetworkSyncManager.cpp
@@ -1599,7 +1599,7 @@ ETTProtocol::Send(const std::string& str)
 			}
 			default: {
 				Locator::getLogger()->warn("NSMAN got unexpected CURLE: {}",
-										   result);
+										   static_cast<size_t>(result));
 				return;
 			}
 		}
